Add move list to puzzle nvui
